requirements.txt
cohere
faiss-cpu
numpy
pdfplumber
python-dotenv
